Move under RBS::Resolver module
