+++ lib/rbs.rb
@@ -28,7 +28,8 @@ require "rbs/definition_builder/method_builder"
 require "rbs/variance_calculator"
 require "rbs/substitution"
 require "rbs/constant"
-require "rbs/constant_resolver"
+require "rbs/resolver/constant_resolver"
+require "rbs/resolver/type_name_resolver"
 require "rbs/constant_table"
 require "rbs/ast/comment"
 require "rbs/writer"
